Removed old activitycard
